requirements.txt
Flask
Flask-Cors
gunicorn
PyJWT
bcrypt
